Fix Mac build, moving linux specific code to its own files
