+++ tools/istio-iptables/pkg/capture/run.go
@@ -21,7 +21,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/containernetworking/plugins/pkg/ns"
 	"github.com/vishvananda/netlink"
 
 	"istio.io/istio/tools/istio-iptables/pkg/builder"
@@ -276,21 +275,9 @@ func ConfigureRoutes(cfg *config.Config, ext dep.Dependencies) error {
 			return ext.Run(command, constants.CommandConfigureRoutes)
 		}
 
-		nsContainer, err := ns.GetNS(cfg.NetworkNamespace)
-		if err != nil {
+		if err := nsContainerCode(cfg); err != nil {
 			return err
 		}
-		defer nsContainer.Close()
-
-		return nsContainer.Do(func(ns.NetNS) error {
-			if err := configureIPv6Addresses(cfg); err != nil {
-				return err
-			}
-			if err := configureTProxyRoutes(cfg); err != nil {
-				return err
-			}
-			return nil
-		})
 	}
 	// called through 'nsenter -- istio-cni configure-routes'
 	if err := configureIPv6Addresses(cfg); err != nil {
