+++ tools/istio-iptables/pkg/dependencies/implementation_unspecified.go
@@ -7,30 +7,25 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//	http://www.apache.org/licenses/LICENSE-2.0
+//     http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package capture
 
-import (
-	"errors"
+package dependencies
 
-	"istio.io/istio/tools/istio-iptables/pkg/config"
-)
+import "errors"
 
 // ErrNotImplemented is returned when a requested feature is not implemented.
 var ErrNotImplemented = errors.New("not implemented")
 
-// configureTProxyRoutes configures ip firewall rules to enable TPROXY support.
-// See https://www.envoyproxy.io/docs/envoy/latest/configuration/http/http_filters/original_src_filter
-func configureTProxyRoutes(cfg *config.Config) error {
+func (r *RealDependencies) execute(cmd string, ignoreErrors bool, args ...string) error {
 	return ErrNotImplemented
 }
 
-func nsContainerCode(cfg *config.Config) error {
+func (r *RealDependencies) executeXTables(cmd string, ignoreErrors bool, args ...string) error {
 	return ErrNotImplemented
 }
